requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
tiktoken>=0.5.0
